+++ lib/tapioca/dsl/compilers/active_record_relation_types.rb
@@ -56,7 +56,7 @@ module Tapioca
         class << self
           extend T::Sig
 
-          sig { override.returns(T::Enumerable[Module]) }
+          sig { override.returns(T::Enumerable[T::Module[T.anything]]) }
           def gather_constants
             Tapioca::Dsl::Compilers::ActiveRecordRelations.gather_constants
           end